requirements.txt
requests>=2.31.0
pandas>=2.1.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
gspread>=6.1.0
